modify control

+++ backend/controllers/Publication.js
@@ -6,18 +6,34 @@ exports.createPost =
      if (!publiBody.content) {
           return res.status().json( { message: 'votre post ne peut pas être vide' } )
      } else {
+          console.log(req.auth)
           const publication = new db.Publication({
+               userId: req.auth.userId,
                ...publiBody
           });
           publication.save()
           .then(() => res.status(201).json({ message: 'post enregistré !'}))
-          .catch(error => res.status(400).json({ error }));
+          .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
      }
    };
 
 exports.modifyPost =
 (req, res, next) => {
-     res.status(200).json({message: 'test modify post'});
+     db.Publication.findOne({where: { _id: req.params.id}})
+     .then(pub => {
+          if(pub.userId !== req.auth.userId){
+               return res.status(403).json ( {message: 'cette requête n\'est pas autorisé'} )
+          };
+          if(req.body.likes || req.body.usersLiked){
+               return res.status(403).json ( {message: 'cette requête n\'est pas autorisé'} )
+          };
+          pub = Object.assign( pub, req.body)
+          .then( newPub => {
+               newPub.save()
+          })
+          .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
+     })
+     .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
 };
 
 exports.deletePost =
@@ -27,25 +43,22 @@ exports.deletePost =
           if(!publication) {
                return res.status(403).json ( {message: 'ce post n\'existe pas'} )
           }
-          publication.destroy({ _id: req.body._id })
+          if (req.auth.userId !== publication.userId) { //vérifie l'identité de l'utilisateur
+               return res.status(403).json ( {message: 'cette requête n\'est pas autorisé'} )
+          } else {
+               publication.destroy({ _id: req.body._id })
                .then(() => res.status(200).json({ message: 'post supprimé !'}))
-               .catch(error => res.status(400).json({ error }));
-          // if (req.auth.userId !== publication.userId) { //vérifie l'identité de l'utilisateur
-          //      return res.status(403).json ( {message: 'cette requête n\'est pas autorisé'} )
-          // } else {
-          //      publication.destroy({ _id: req.body._id })
-          //      .then(() => res.status(200).json({ message: 'post supprimé !'}))
-          //      .catch(error => res.status(400).json({ error }));
-          // }
+               .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
+          }
      })
-     .catch(error => res.status(500).json({ error }));
+     .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
 };
 
 exports.getAllPost =
 (req, res, next) => {
      db.Publication.findAll()
      .then(publications=> res.status(200).json(publications))
-     .catch(error => res.status(400).json({ error }));
+     .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
    };
 
 exports.getOnePost =
@@ -58,13 +71,13 @@ exports.getOnePost =
                res.status(200).json(publication);
           };
      })
-     .catch(error => res.status(404).json({ error }));
+     .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
 };
 
 exports.likePost =
 (req, res, next) => {
      const like = req.body.like;
-     const userId = req.body.userId;
+     const userId = req.auth.userId;
      if(like > 1 || like < 0){
           return res.status(401).json( { message: 'cette requête n\'est pas autorisé' } );
      };
@@ -75,6 +88,10 @@ exports.likePost =
           } else {
                db.Publication.findOne( { where: { _id: req.params._id } } )
                .then(pub =>{
+                    if(pub.userId === req.auth.userId){
+                         return res.status(403).json( { message: 'Vous ne pouvez pas liker votre propre post' } )
+                    };
+
                     const usersLiked = JSON.parse(pub.usersLiked);
 
                     if(like === 1){
@@ -86,7 +103,7 @@ exports.likePost =
                          }
                          pub.save()
                          .then ( res.status(201).json( { message: 'like ajouté' } ))
-                         .catch(error => res.status(404).json({ error }));
+                         .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
                     }
 
                     if(like === 0){
@@ -98,11 +115,11 @@ exports.likePost =
                          }
                          pub.save()
                          .then ( res.status(201).json( { message: 'like enlevé' } ))
-                         .catch(error => res.status(404).json({ error }));
+                         .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
                     }
                })
-               .catch(error => res.status(404).json({ error }));
+               .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
           };
      })
-     .catch(error => res.status(404).json({ error }));
+     .catch(error => res.status(500).json({ message: `oops! something went wrong... ${error}` }));
 };
